## 2018-06-10 System Themes

+++ Source/Krypton Components/ComponentFactory.Krypton.Navigator/Dragging/DropDockingIndicatorsSquare.cs
@@ -110,7 +110,7 @@ namespace ComponentFactory.Krypton.Navigator
             }
 
 		    // Show the window without activating it (i.e. do not take focus)
-			PI.ShowWindow(Handle, (short)PI.SW_SHOWNOACTIVATE);
+			PI.ShowWindow(Handle, PI.SW_SHOWNOACTIVATE);
 		}
 
 		/// <summary>
